Package All Files Respectively
Currently all files are under the source root file.

Lets,
* Move all files into their respective packages.
